Add delete booking functionality
Currently calendar refetches the entire event source on a booking
operation. We do this to ensure the booking information, for the
event, is actually updated (preventing need to refresh page...).
I am currently researching solution to this, that is finding a
way to refetch/update the changed event only. Additionally, a
placeholder uid/family_id (1,1) is used for adding/removing
a booking. This needs to be updated to reflect the actual
uid & family id of the user requesting the action. However,
for the demo on tuesday, the current functionality will be more
than sufficient, in my opinion.